requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
azure-servicebus>=7.11.1
azure-identity>=1.13.0
pydantic>=2.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3
typing-extensions>=4.5.0 